requirements.txt
appdirs==1.4.4
asgiref==3.2.10
backcall==0.2.0
black==21.5b1
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
coreapi==2.3.3
coreschema==0.0.4
decorator==5.0.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1
django-import-export==2.5.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
idna==2.10
ipython==7.23.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib-inline==0.1.2
mypy-extensions==0.4.3
numpy==1.20.3
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.2.4
parso==0.8.2
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycountry==20.7.3
Pygments==2.9.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.4
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
